test: separate tests to prevent port collision
